Better Rest API for CategoryController

+++ my_app_rest_api/src/main/java/com/gruszm/kontenery/projekt_konteneryzacja/controllers/CategoryController.java
@@ -1,8 +1,11 @@
 package com.gruszm.kontenery.projekt_konteneryzacja.controllers;
 
 import com.gruszm.kontenery.entities.*;
+import com.gruszm.kontenery.projekt_konteneryzacja.exceptions.CategoryAlreadyExistsException;
 import com.gruszm.kontenery.projekt_konteneryzacja.services.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -21,32 +24,69 @@ public class CategoryController
     }
 
     @GetMapping
-    public List<Category> getAllCategories()
+    public ResponseEntity<List<Category>> getAllCategories()
     {
-        return categoryService.findAll();
+        return ResponseEntity.status(HttpStatus.OK).body(categoryService.findAll());
     }
 
     @GetMapping("/name/{name}")
-    public Category getCategoryByName(@PathVariable(name = "name") String name)
+    public ResponseEntity<Category> getCategoryByName(@PathVariable(name = "name") String name)
     {
-        return categoryService.findByName(name);
+        Category category = categoryService.findByName(name);
+
+        if (category == null)
+        {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
+        else
+        {
+            return ResponseEntity.status(HttpStatus.OK).body(category);
+        }
     }
 
     @PostMapping
-    public void saveCategory(@RequestBody Category category)
+    public ResponseEntity<Category> saveCategory(@RequestBody Category category)
     {
-        categoryService.save(category);
+        try
+        {
+            categoryService.save(category);
+            return ResponseEntity.status(HttpStatus.OK).body(category);
+        }
+        catch (CategoryAlreadyExistsException e)
+        {
+            return ResponseEntity.status(HttpStatus.CONFLICT).body(null);
+        }
     }
 
     @DeleteMapping("/id/{id}")
-    public void deleteCategoryById(@PathVariable String id)
+    public ResponseEntity<Category> deleteCategoryById(@PathVariable String id)
     {
-        categoryService.deleteById(id);
+        Category category = categoryService.findById(id);
+
+        if (category == null)
+        {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
+        else
+        {
+            categoryService.deleteById(id);
+            return ResponseEntity.status(HttpStatus.OK).body(category);
+        }
     }
 
     @DeleteMapping("/name/{name}")
-    public void deleteCategoryByName(@PathVariable String name)
+    public ResponseEntity<Category> deleteCategoryByName(@PathVariable String name)
     {
-        categoryService.deleteByName(name);
+        Category category = categoryService.findByName(name);
+
+        if (category == null)
+        {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
+        else
+        {
+            categoryService.deleteByName(name);
+            return ResponseEntity.status(HttpStatus.OK).body(category);
+        }
     }
 }
